Move files into tsc module
lua/tsc.lua -> lua/tsc/init.lua

lua/utils.lua -> lua/tsc/utils.lua
