+++ lua/tsc/init.lua
@@ -1,5 +1,5 @@
 local success, pcall_result = pcall(require, "notify")
-local utils = require("utils")
+local utils = require("tsc.utils")
 
 local M = {}
